% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_WebScraping_DBMaddison.R
\name{get.DBMaddison}
\alias{get.DBMaddison}
\title{Maddison Project Database 2020}
\usage{
get.DBMaddison(
  country = as.character(),
  start = as.numeric(),
  end = as.numeric()
)
}
\arguments{
\item{country}{representa o nome do país desejado. O parâmetro aceita mais de um nome por vez ou, até mesmo, o nome de todas as séries simultaneamente passando no parâmetro country = \code{"all"};}

\item{start}{representa o início do período/ano selecionado}

\item{rend}{representa o fim do período/ano selecionado}
}
\value{
Retorna um data.frame com os series selecionadas
}
\description{
Maddison Project Database 2020 foi coletado via Web Sraping. As séries disponíveis na função \code{get.DBMaddison} são as seguintes:

\code{gdpcc} e
\code{pop}.

São respestivamente: Real GDP per capita in 2011$ e Population - mid-year (thousands). Há dados disponíveis desde do século XIX ou até mesmo de períodos anteriores (Para o Brasil em específico há dados desde 1800). Para mais detalhes sobre a metodologia de estimação/retropolação consultar a documentação no Github.
}
\details{
O Maddison Project Database fornece informações sobre crescimento econômico comparativo e níveis de renda no longo prazo. A versão 2020 deste banco de dados abrange 169 países e o período até 2018.
}
\examples{
\dontrun{


# Maddison Project Database 2020

teste_mesa_all = CliometricsBR::get.DBMaddison(country = "all",
                                               start = 1800, end = 1900)

teste_mesa_filter = CliometricsBR::get.DBMaddison(country = c("USA", "BRA", "ARG"),
                                                  start = 1850, end = 1973)

teste_mesa_BRA= CliometricsBR::get.DBMaddison(country = "BRA")

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
