% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_censo1872.R
\name{get_Censo1872}
\alias{get_Censo1872}
\alias{get_censo1872}
\title{Dados - Recenseamento de 1872}
\usage{
get_censo1872()
}
\description{
Dados - Recenseamento de 1872
}
\examples{
\dontrun{


}


}
\author{
Luiz Paulo Tavares Gonçalves
}
