% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_exports.R
\name{get_exports}
\alias{get_exports}
\title{Dados - Exportações}
\usage{
get_exports(series = as.character(), start = as.numeric(), end = as.numeric())
}
\arguments{
\item{series}{Representa o nome da série desejada. O parâmetro aceita mais de um nome por vez ou, até mesmo, o nome de todas as séries simultaneamente passando no parâmetro series = \code{"all"};}

\item{start}{Representa o ano de início da série que deseja}

\item{end}{Por sua vez, \code{end}, representa o último ano da série que deseja, isto é, o fim do período desejado.}
}
\value{
Retorna um data.frame com os series selecionadas
}
\description{
As séries disponíveis na função \code{get_exports} são as seguintes:

\code{Yerba_mate};
\code{Sugar};
\code{Cotton};
\code{Cocoa};
\code{Tobacco};
\code{Coffee} e
\code{Fob}.

Respectivamente, exportações de: Erva-mate, Açúcar, Algodão, Cacau, Tabaco, Café e exportações em FOB (Free on Board ou Livre a Bordo do Navio).
}
\examples{
\dontrun{

exports_all = CliometricsBR::get_exports(series = "all",
                                        start = 1821, end = 1900)

# Importando produtos selecionados = Sugar, Coffee e Cotton
# De 1821 a 1855

exports_select = CliometricsBR::get_exports(series = c("Sugar", "Coffee", "Cotton"),
                                           start = 1821, end = 1855)

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
