% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prices.R
\name{get_prices}
\alias{get_prices}
\title{Dados - Cotações/Preço}
\usage{
get_prices(units = as.character())
}
\arguments{
\item{units}{representa as unidades disponíveis: Bag (em sacas de 50kg) ou em Kilo (quilo)}
}
\value{
Retorna um data.frame com as unidades selecionadas
}
\description{
A função retorna o preço médio do café - de 1808 a 1900 (ou seja, pegando o período anterior ao Brasil Império até a República)

Estão disponíveis na função o preço em quilo ou em sacas de 50kg
}
\details{
A unidade monetária das sacas de café estão em Franco francês, por sua vez, o quilo está em Réis.
}
\examples{
\dontrun{

  teste_prices_all = CliometricsBR::get_prices(units = "all")

  teste_prices_select = CliometricsBR::get_prices(units = "Kilo")

  teste_prices= CliometricsBR::get_prices(units = c("Kilo", "Bag"))

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
