% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_series.R
\name{get_seriesIPEA}
\alias{get_seriesIPEA}
\title{Séries históricas disponíveis no IPEAdata}
\usage{
get_seriesIPEA(codes = as.character())
}
\arguments{
\item{codes}{É o parâmetro que recebe o código da série}
}
\value{
Retorna um data.frame as séries históricas selecionadas
}
\description{
A função busca retorna todas as séries históricas disponíveis no IPEAdata via API
}
\examples{
\dontrun{

  series_all = CliometricsBR::get_seriesIPEA(codes = c("HIST_ACOV",
                                                       "HIST_ACUCARQ",
                                                       "HIST_ACUCARV"))

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
