% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_db}
\alias{load_db}
\title{Base de dados do pacote CliometricsBR}
\usage{
load_db()
}
\value{
Retorna um data.frame com a base de dados bruta utilizada no pacote
}
\description{
A função representa a importação da base de dados utilizada no pacote
}
\examples{
\dontrun{

         db_cliometrics = load_db()
}


}
\author{
Luiz Paulo Tavares Gonçalves
}
