% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipea_metadados.R
\name{get_metadadosIPEA}
\alias{get_metadadosIPEA}
\title{Séries históricas disponíveis no IPEAdata}
\usage{
get_metadadosIPEA()
}
\value{
Retorna um data.frame com todas as séries históricas disponíveis
}
\description{
A função busca retorna todas as séries históricas disponíveis no IPEAdata via API
}
\details{
Não há parâmetro na função, basta chamar a função que irá retornar todas as séries históricas disponíveis
}
\examples{
\dontrun{

  series_available = CliometricsBR::get_metadadosIPEA()

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
