% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_db.R
\name{load_db}
\alias{load_db}
\title{Base de dados do pacote CliometricsBR}
\usage{
load_db()
}
\value{
Retorna um data.frame com a base de dados bruta utilizada no pacote, com exceção dos dados de Web-Scraping.
}
\description{
A função representa a importação da base de dados local (pasta {R}) utilizada no pacote. Portanto, os dados utilizados no CliometricsBR que são derivados de Web-Scraping não estão disponíveis no presente banco de dados local. Assim, possibilitando um melhor gerenciamento do pacote.
}
\examples{
\dontrun{

         db_cliometrics = load_db()
}


}
\author{
Luiz Paulo Tavares Gonçalves
}
